Binary Trees Program
Objectives:
Practice implementing a binary search tree.
Practice coming up with a functional breakdown for a large program.
Practice efficiently updating all necessary components of a data structure after each update

Problem:
Write a program that reads in input corresponding to various changes and queries to the theater's
loyalty program and prints out corresponding messages for each of the input commands. Here is
an informal list of each of the possible commands in the input:
(1) Add Loyalty Points to a particular customer.
(2) Subtract Loyalty Points from a particular customer.
(3) Delete a particular customer.
(4) Search for a particular customer in the binary search tree. If the customer is found, report
both their number of loyalty points and their depth in the tree (distance from the root in # of links
to follow.)
(5) Count the number of customers whose names come alphabetically before a particular
customer.
At the very end of the input, your program should store pointers to each struct storing customer
data and sort that data by loyalty points, from highest to lowest, breaking ties alphabetically. (For
two customers with the same number of loyalty points, the one whose name comes first
alphabetically should be listed first.) This data should be sorted via either Merge Sort or Quick
Sort.